refactor: Refactor CustomNode Modal
