+++ src/pages/flow-canvas/styles/BodyEditor.module.scss
@@ -116,4 +116,13 @@
   padding: 0;
   border: 1px solid $primary-700;
   border-radius: $radius-sm;
+  outline: none;
+}
+
+.editableValue {
+  cursor: pointer;
+  padding: 2px 4px;
+  &:hover {
+    background-color: $white;
+  }
 }
